move NodeInfo into runtimecontext pkg
