+++ java/api/src/main/java/org/ray/api/Ray.java
@@ -4,6 +4,7 @@ import java.util.List;
 import org.ray.api.id.UniqueId;
 import org.ray.api.runtime.RayRuntime;
 import org.ray.api.runtime.RayRuntimeFactory;
+import org.ray.api.runtimecontext.RuntimeContext;
 
 /**
  * This class contains all public APIs of Ray.
