+++ java/api/src/main/java/org/ray/api/runtimecontext/NodeInfo.java
@@ -1,4 +1,4 @@
-package org.ray.api;
+package org.ray.api.runtimecontext;
 
 import java.util.Map;
 import org.ray.api.id.UniqueId;
