+++ java/runtime/src/main/java/org/ray/runtime/RuntimeContextImpl.java
@@ -2,9 +2,9 @@ package org.ray.runtime;
 
 import com.google.common.base.Preconditions;
 import java.util.List;
-import org.ray.api.NodeInfo;
-import org.ray.api.RuntimeContext;
 import org.ray.api.id.UniqueId;
+import org.ray.api.runtimecontext.NodeInfo;
+import org.ray.api.runtimecontext.RuntimeContext;
 import org.ray.runtime.config.RunMode;
 import org.ray.runtime.task.TaskSpec;
