+++ java/runtime/src/main/java/org/ray/runtime/gcs/GcsClient.java
@@ -9,8 +9,8 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.ArrayUtils;
 import org.ray.api.Checkpointable.Checkpoint;
-import org.ray.api.NodeInfo;
 import org.ray.api.id.UniqueId;
+import org.ray.api.runtimecontext.NodeInfo;
 import org.ray.runtime.generated.ActorCheckpointIdData;
 import org.ray.runtime.generated.ClientTableData;
 import org.ray.runtime.generated.TablePrefix;
